+++ Tools/imgCIF_app/tui.py
@@ -20,6 +20,7 @@ from .core import (
     guess_file_type,
     make_cif,
 )
+from .helpers import extrapolate_sequence, guess_doi
 
 # Recognised formats for ImgCIF
 ARCHIVE_TYPES = ("TGZ", "TBZ", "TXZ", "ZIP")
@@ -66,35 +67,6 @@ class DOIValidator(Validator):
                     message=f"Bad DOI ({resp.status_code}: {resp.reason})",
                     cursor_position=cp,
                 )
-
-
-DOI_RULES = [
-    # Download URL regex -> DOI template
-    (r"https://zenodo\.org/records/(\d+)", "10.5281/zenodo.{}"),
-    (r"\w+://[\w\-.]+/10\.15785/SBGRID/(\d+)", "10.15785/SBGRID/{}"),  # Various sbgrid domains
-    (r"https://xrda\.pdbj\.org/rest/public/entries/download/(\d+)", "10.51093/xrd-{:05}"),
-]
-
-
-def guess_doi(download_info):
-    urls = []
-    for loc in download_info:
-        if isinstance(loc, ArchiveUrl):
-            urls.append(loc.url)
-        elif isinstance(loc, DirectoryUrl):
-            urls.append(loc.url_base)
-
-    if not urls:
-        return ""
-
-    for url_pat, doi_template in DOI_RULES:
-        matches = [re.match(url_pat, u) for u in urls]
-        if all(matches):
-            id_part = matches[0][1]
-            if all(m[1] == id_part for m in matches[1:]):
-                return doi_template.format(id_part)
-
-    return ""
 
 
 def check_url(url, msg="Checking URL..."):
@@ -147,35 +119,6 @@ def find_common_ancestor(p1: Path, p2: Path):
         if p2.is_relative_to(candidate):
             return candidate
     raise ValueError(f"No ancestor in common: {p1} & {p2}")
-
-
-def extrapolate_sequence(s0, s1, length):
-    matched0 = re.split(r"(\d+)", s0)
-    matched1 = re.split(r"(\d+)", s1)
-    if len(matched0) != len(matched1):
-        return
-
-    if (
-        len(
-            diffs := [
-                i for i, (p0, p1) in enumerate(zip(matched0, matched1)) if p0 != p1
-            ]
-        )
-        != 1
-    ):
-        return  # No difference, or >1 piece differs
-    if (diff_ix := diffs[0]) % 2 == 0:
-        return  # The difference is in a non-numeric part
-
-    width = len(matched0[diff_ix])
-    n0 = int(matched0[diff_ix])  # First number in sequence
-    if int(matched1[diff_ix]) != n0 + 1:
-        return  # Not increasing by 1
-
-    for i in range(2, length):
-        pieces = matched0.copy()
-        pieces[diff_ix] = f"{n0 + i:0{width}}"
-        yield "".join(pieces)
 
 
 def get_download_urls(expts: ExperimentList):
